Allow banner config in site config (#22)
As-is, `usabanner` must be configured in each page's params (as does
`draftbanner`) due to using `.Params.usabanner`. Using the `.Param`
function instead allows looking up first at the page level and then in
the site params. This will allow for enabling the banner globally and
then (optionally) disabling it on select pages.