Use fingerprint in stylesheet/script links
Effectively, this enables a form of cache-busting for these files as the
hash for the file is included in the URL when following this scheme.
This was already done for one of the files (on line 65, the reference to
`pageCSSminSecure`). Because SHA-512 is in use and the string isn't
trimmed, this results in fairly long file names (but again, that was
already being done in at least one place). This should make it easier
for consumers to deal with fewer caching issues (and needing to
invalidate fewer files on a CDN) after a deployment with style changes.